+++ src/view/View.js
@@ -1,0 +1,28 @@
+
+import {createElement} from '../utils/render';
+
+export default class View {
+  constructor() {
+    if (new.target === View) {
+      throw new Error(`Can't instantiate Abstract, only concrete one.`);
+    }
+    this._element = null;
+    this._callback = {};
+  }
+
+  _getTemplate() {
+    throw new Error(`Abstract method not implemented: getTemplate`);
+  }
+
+  getElement() {
+    if (!this._element) {
+      this._element = createElement(this._getTemplate());
+    }
+    return this._element;
+  }
+
+  removeElement() {
+    this._element.remove();
+    this._element = null;
+  }
+}
