+++ src/view/menu.js
@@ -1,4 +1,4 @@
-import {createElement} from "../utils.js";
+import View from './View';
 
 const createMenuTemplate = () => {
   return (
@@ -12,24 +12,8 @@ const createMenuTemplate = () => {
   );
 };
 
-export default class Menu {
-
-  constructor() {
-    this._element = null;
-  }
+export default class Menu extends View {
   _getTemplate() {
     return createMenuTemplate();
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this._getTemplate());
-    }
-    return this._element;
-  }
-
-  removeElement() {
-    this._element.remove();
-    this._element = null;
   }
 }
